requirements.txt
adk-python==1.16.0
google-genai>=0.8.0
requests==2.31.0
pydantic==2.11.5
uvicorn==0.34.0
psycopg==3.2.4
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
neo4j==5.27.0
fastapi==0.112.2
google-cloud-firestore==2.18.0

